requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
Click==7.0
configparser==3.5.0
cryptography==2.1.4
enum34==1.1.6
Flask==1.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
futures==3.2.0
idna==2.6
ipaddress==1.0.17
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.13.3
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.3
pyxdg==0.25
PyYAML==5.1.1
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.3.4
vboxapi==1.0
virtualenv==16.0.0
Werkzeug==0.15.4
wrapt==1.10.11
